Deeper 0.3 for Windows Released!
Deeper 0.3 is a major update that brings a lot of explanation to the game, to help you understand it better, along with a new colorful UI and advanced World Saving System.

New features:
+Added a "Help" button to give you the basics of how to play the game.
+Added a Tutorial World that guides you step by step on learning to play Deeper.
+Brand new Colorful easy-to-use User Interface.
+Added a new World Saving System, which allows you to save/load multiple worlds, instead of just one.
+Added new features to the in-game menu that shows you what world you are playing on and what Item you are using.

Looking for the OS X Installer? That's currently not ready yet, but it will be up in a couple of days.